% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{compute_flatline_score}
\alias{compute_flatline_score}
\title{Computes the quality scores of the rows in \code{dataset}.}
\usage{
compute_flatline_score(x, id.var = NULL)
}
\arguments{
\item{x}{A dataset that contains every subject (rows) and all variables (columns) to be used. Or a numeric vector to compute a single score.}

\item{id.var}{Either a numeric or character vector identifying a column in the dataset, or a vector of the same length as the amount of rows in x.}
}
\value{
A numeric vector of scores. A named vector in case id.var is non NULL.
}
\description{
Computes the quality scores of the rows in \code{dataset}.
}
\note{
The score is defined as the proportion of scores that has the value for a variable as the variable before it.
}
