% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{computeQualityScore}
\alias{computeQualityScore}
\title{Computes the quality score of the respondent at \code{index} of \code{dataset}. The score is defined as the proportion of scores that has the value for a variable as the variable before it.}
\usage{
computeQualityScore(index, dataset)
}
\arguments{
\item{index}{The index of the respondent in the dataset.}

\item{dataset}{A dataset that contains all variables to base the score on.}
}
\value{
A numeric score
}
\description{
Computes the quality score of the respondent at \code{index} of \code{dataset}. The score is defined as the proportion of scores that has the value for a variable as the variable before it.
}
\note{
This function requires at least two variables in the dataset.
}
