% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{computeQualityScores}
\alias{computeQualityScores}
\title{Computes the quality scores of the respondents in \code{dataset}. The score is defined as the proportion of scores that has the value for a variable as the variable before it.}
\usage{
computeQualityScores(dataset)
}
\arguments{
\item{dataset}{A dataset that contains all variables to base the score on.}
}
\value{
A numeric vector of scores.
}
\description{
Computes the quality scores of the respondents in \code{dataset}. The score is defined as the proportion of scores that has the value for a variable as the variable before it.
}
\note{
This function requires at least two variables in the dataset.
}
