% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_respondent}
\alias{plot_respondent}
\title{Returns a plot for the respondent at \code{index} of \code{data}}
\usage{
plot_respondent(data, title = NULL)
}
\arguments{
\item{data}{A dataset that contains all variables and respondents to be plotted.}

\item{index}{The index of the respondent in the data.}

\item{xlab}{A vector of titles that should be given to the plots. Can be \code{null}. See examples.}
}
\value{
A ggplot
}
\description{
Returns a plot for the respondent at \code{index} of \code{data}
}
