% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_respondents}
\alias{plot_respondents}
\title{Return plots for all respondents in \code{x}}
\usage{
plot_respondents(x, id.var = NULL)
}
\value{
A ggplot
}
\description{
Return plots for all respondents in \code{x}
}
