% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQ.R
\name{plotRespondents}
\alias{plotRespondents}
\title{Return plots for all respondents in \code{data}}
\usage{
plotRespondents(data, titles = NULL)
}
\arguments{
\item{data}{A dataset that contains all variables and respondents to be plotted.}

\item{titles}{A vector of titles that should be given to the plots. Can be \code{null}. See examples.}
}
\value{
A ggplot
}
\description{
Return plots for all respondents in \code{data}
}
\examples{
ids <- paste0("ID_",1:10)
data <- data.frame(ID = ids, L = sample(1:7, 10, replace = T), P = sample(1:7, 10, replace = T))
plots <- plotRespondents(data = mpg, titles = ids)

}
