% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQ.R
\name{plotRespondent}
\alias{plotRespondent}
\title{Returns a plot for the respondent at \code{index} of \code{data}}
\usage{
plotRespondent(index, data, title = index)
}
\arguments{
\item{index}{The index of the respondent in the data.}

\item{data}{A dataset that contains all variables and respondents to be plotted.}

\item{xlab}{A vector of titles that should be given to the plots. Can be \code{null}. See examples.}
}
\value{
A ggplot
}
\description{
Returns a plot for the respondent at \code{index} of \code{data}
}
\examples{
ids <- paste0("ID_",1:10)
data <- data.frame(ID = ids, L = sample(1:7, 10, replace = T), P = sample(1:7, 10, replace = T))
plot <- plotRespondent(index = 1, data = mpg, title = ids[1])

}
