% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRespondents}
\alias{plotRespondents}
\title{Return plots for all respondents in \code{data}}
\usage{
plotRespondents(data, id.var = NULL)
}
\arguments{
\item{data}{A dataset that contains all variables and respondents to be plotted.}

\item{titles}{A vector of titles that should be given to the plots. Can be \code{null}. See examples.}
}
\value{
A ggplot
}
\description{
Return plots for all respondents in \code{data}
}
